# Conflicts:
#	PROJECT-Rogue/Assets/Scenes/Game.unity
#	PROJECT-Rogue/Assets/Scenes/ItemTesting.unity
#	PROJECT-Rogue/Assets/Scripts/Characters/Enemy.cs
#	PROJECT-Rogue/Assets/Scripts/Characters/Player.cs
#	PROJECT-Rogue/Assets/Scripts/Weapons/Bullet.cs
#	PROJECT-Rogue/Assets/Scripts/Weapons/ProjectileRifle.cs
#	PROJECT-Rogue/Assets/Scripts/Weapons/ProjectileSniperRifle.cs
#	PROJECT-Rogue/Assets/Scripts/Weapons/Raycast.cs
#	PROJECT-Rogue/ProjectSettings/TagManager.asset

+++ PROJECT-Rogue/Assets/Scripts/Weapons/Bullet.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Bullet : MonoBehaviour
 {
     public GameObject impactAnimation;
-    float damage;
+    int damage;
     Vector2 velocity;
     string attackerTag;
     private void Update()
