+++ PROJECT-Rogue/Assets/Scripts/Weapons/Raycast.cs
@@ -7,5 +7,5 @@ public abstract class Raycast : Ranged
     LineRenderer lineRenderer;
     protected override void Shoot() { whoAttacks.StartCoroutine(ShootRaycast()); }
     protected abstract IEnumerator ShootRaycast();
-    public Raycast(float attackSpeed = 10, float damage = 20) : base(attackSpeed, damage) { }
+    public Raycast(float attackSpeed = 10, int damage = 20) : base(attackSpeed, damage) { }
 }
